requirements.txt
addict==2.4.0
appdirs==1.4.4
asttokens==2.0.5
attrs==22.2.0
backcall==0.2.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
ConfigArgParse==1.5.3
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
dash==2.8.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.38.0
hole-filling-liepa==0.0.3
idna==3.4
igl==2.2.1
imageio==2.25.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
meshio==5.3.4
mistune==2.0.4
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.0
numpy==1.24.1
open3d==0.16.0
opencv-contrib-python==4.7.0.68
packaging==22.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==2.5.2
plotly==5.13.0
pooch==1.6.0
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyglet==1.5.27
Pygments==2.11.2
pymeshfix==0.16.2
pymeshlab==2022.2.post2
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pyvista==0.37.0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.2.0
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
scooby==0.7.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==65.6.3
SimpleITK-SimpleElastix==2.0.0rc2.dev910
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tenacity==8.1.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.1.23.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trimesh==3.18.1
uri-template==1.2.0
urllib3==1.26.14
vtk==9.2.5
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.5
